requirements.txt
django-tinymce4-lite
